Serial debug printing device.
The "serial-dbg" option causes VxT to create a debug printing device on the
serial port. Each character transmitted to the COM1 port by the guest is
printed to standard output on the host.

This option replaces the serial mouse; when it's enabled, serial_dbg.c is
compiled into libvxt and mouse.c is not.